Move nvim config to saved nvim-plain.
Started to use AstroNvim to simplify configuration of nvim so saving
away my existing nvim configuration (instead of deleting).

AstroNvim uses the .config/nvim directory for the base of the framework,
so .config/nvim is now the AstroNvim framework under their projects git
control and is not part of this repo any longer.